# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CameraIn": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "resolution_height": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 4320.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Height"
          },
          "resolution_width": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 7680.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Width"
          },
          "room_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "CameraIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ONE API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/homes/{home_id}/cameras": {
      "get": {
        "operationId": "cameras_api_v1_homes__home_id__cameras_get",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cameras"
      },
      "post": {
        "operationId": "camera_api_v1_homes__home_id__cameras_post",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Camera"
      }
    },
...
  }
}